fix(skia): use committed wasm libs on non-building hosts, never remote-fetch
The compiled lib/*.wasm are committed to the repo now (CI rebuilds and commits
them on skia changes), so the old remote-prebuilt fetch would overwrite the
tracked libs with a stale published version and dirty git history. Drop it.

On a host that can't compile (macOS 27 / ziglang#31658), setup.mjs and
build-wasm.mjs now use the committed libs when present and fail hard when
they're missing — never fetch. Remove the dead skia:fetch-wasm script and its
stale prebuilt-wasm.json manifest; rename prebuilt-wasm.mjs to host-capability.mjs
since only the canBuildWasm host probe survives.
