+++ packages/skia/package.json
@@ -113,7 +113,6 @@
     "skia:check": "node scripts/setup.mjs --check",
     "skia:tools": "node scripts/setup.mjs --tools",
     "build": "node scripts/setup.mjs --ensure && tsdown",
-    "skia:fetch-wasm": "node scripts/prebuilt-wasm.mjs",
     "build:wasm:gl": "node scripts/build-wasm.mjs --gl-only",
     "build:wasm:wgpu": "node scripts/build-wasm.mjs --wgpu-only",
     "dev": "tsdown --watch",
