--- packages/skia/prebuilt-wasm.json
@@ -1,9 +1,0 @@
-{
-  "_comment": "Prebuilt WASM published to npm, used as a fallback when the local host cannot build the Zig->WASM toolchain (e.g. macOS 26.4+/27 dropped the arm64 libSystem slice, ziglang/zig#31658). CI/Linux builds from source and remains the source of truth. When a new WASM is published, bump `version` and update the sha256 hashes (sha256 of package/dist/<artifact> in the published tarball).",
-  "package": "@three-flatland/skia",
-  "version": "0.1.0-alpha.3",
-  "artifacts": {
-    "lib/skia-gl.wasm": "sha256:e7c9ad7f785e6814143f8f0476407141e4772e06a3fe35f1eb1347ebe565946d",
-    "lib/skia-wgpu.wasm": "sha256:8cbe0d7f34d2e01debb57f97c54e3fec56fd4a17b0313c9923d649eabea0395e"
-  }
-}
